On Mac, don't cache the results of C compiles
When I upgraded Xcode on my Mac, some of my Bazel builds that were using
the remote cache stopped working. It turns out it is because the C
compiler and header files updated, but Bazel does not store all the
appropriate information about those files when it uploads .o and .d
files to the remote cache.

Workaround: Just don't cache .o and .d files. (Or specifically, .pic.o
and .pic.d files)

https://app.asana.com/0/215550984114637/306079885945759