Remove blacklisting of .o files for remote cache
Previously, we blacklisted files with extension ".o" and ".d" when
compiling on OSX, to prevent them from being uploaded to or downloaded
from our remote cache. But in order to reduce the number of Bazel
customizations we have, we have now changed our Bazel project to prevent
the CppCompile action from using strategy "remote", so now we can remove
this customization.